move `custom-sharp.js` into separate test fixture
