+++ test/integration/image-optimizer/test/util.ts
@@ -147,26 +147,6 @@ export function runTests(ctx) {
     slowImageServer.stop()
   })
 
-  if (!isDev && ctx.isSharp && ctx.nextConfigImages) {
-    it('should handle custom sharp usage', async () => {
-      const res = await fetchViaHTTP(ctx.appPort, '/api/custom-sharp')
-
-      expect(res.status).toBe(200)
-      expect(await res.json()).toEqual({ success: true })
-      const traceFile = await fs.readJson(
-        join(
-          ctx.appDir,
-          '.next',
-          'server',
-          'pages',
-          'api',
-          'custom-sharp.js.nft.json'
-        )
-      )
-      expect(traceFile.files.some((file) => file.includes('sharp/'))).toBe(true)
-    })
-  }
-
   if (domains.length > 0) {
     it('should normalize invalid status codes', async () => {
       const url = `http://localhost:${
@@ -1430,20 +1410,6 @@ export const setupTests = (ctx) => {
         })
         curCtx.nextOutput = ''
         nextConfig.replace('{ /* replaceme */ }', json)
-
-        if (curCtx.isSharp) {
-          await fs.writeFile(
-            join(curCtx.appDir, 'pages', 'api', 'custom-sharp.js'),
-            `
-            import sharp from 'sharp'
-            export default function handler(req, res) {
-              console.log(sharp)
-              res.json({ success: true })
-            }
-          `
-          )
-        }
-
         await nextBuild(curCtx.appDir)
         await cleanImagesDir(ctx)
         curCtx.appPort = await findPort()
@@ -1461,11 +1427,6 @@ export const setupTests = (ctx) => {
       })
       afterAll(async () => {
         nextConfig.restore()
-        if (curCtx.isSharp) {
-          await fs.remove(
-            join(curCtx.appDir, 'pages', 'api', 'custom-sharp.js')
-          )
-        }
         if (curCtx.app) await killApp(curCtx.app)
       })
