+++ test/integration/sharp-api/app/package.json
@@ -1,0 +1,7 @@
+{
+  "private": true,
+  "packageManager": "npm@10.4.0",
+  "dependencies": {
+    "sharp": "latest"
+  }
+}
