+++ test/integration/sharp-api/app/pages/api/custom-sharp.js
@@ -1,0 +1,5 @@
+import sharp from 'sharp'
+
+export default function handler(req, res) {
+  res.json({ success: Boolean(sharp) })
+}
